chore: 마지막 커서 기억 함수 이동
